+++ src/hotspot/share/utilities/decoder_elf.cpp
@@ -67,7 +67,7 @@ bool ElfDecoder::get_source_info(address pc, char* filename, size_t filename_len
   char filepath[JVM_MAXPATHLEN];
   filepath[JVM_MAXPATHLEN - 1] = '\0';
   int offset_in_library = -1;
-  if (!os::dll_address_to_library_name(pc, filepath, sizeof(filepath), &offset_in_library)) {
+  if (!os::dll_address_to_library_name(pc, filepath, (int)sizeof(filepath), &offset_in_library)) {
     // Method not found. offset_in_library should not overflow.
     DWARF_LOG_ERROR("Did not find library for address " PTR_FORMAT, p2i(pc))
     return false;
